Separate `services` module in database
