+++ uds/database/__init__.py
@@ -5,4 +5,4 @@ Tools for decoding and encoding information from/to diagnostic messages.
 """
 
 from .abstract_data_record import AbstractDataRecord, DataRecordType, DecodedDataRecord
-from .abstract_service import AbstractService
+from .services import AbstractService
