+++ uds/database/services/__init__.py
@@ -1,0 +1,3 @@
+"""Database implementation for diagnostic services."""
+
+from .abstract_service import AbstractService
